requirements.txt
# SmartAPI dependencies
smartapi-python==1.5.2
pyotp>=2.8.0
websocket-client>=1.8.0
logzero>=1.7.0
requests>=2.31.0
pycryptodome>=3.22.0

# Data processing
pandas>=1.3.0
numpy>=1.20.0
pytz>=2023.3

# ClickHouse
clickhouse-driver>=0.2.4

# Utilities
python-dateutil>=2.8.2